Passthrough shaders when using Vulkan
We should maybe provide a flag that could force to use validation, since that is useful when creating the shaders.

This has the nice benefit, that WGPU won't touch the SPIR-V and use them directly, so loading is a bit faster and maybe even the shader might be optimized better by the driver.