Replace resampling with rubato
But we vendor Rubato, which is also MIT code. I updated the MIT license file of korangar_audio to respect the author.

Why the vendoring? It's mainly to reduce dependencies, make the API easier to understand and most importantly, changing the window function of the resampler.

The original window function would "darken" 22050 Hz audio files when resampling them, since BlackmanHarris2 has a pretty wide transition bandwidth, which goes into the audible range.

We instead use Kaiser β=10, which has a much lower transition bandwidth but still enough stopband attenuation for 16-bit audio.

This implementation also caches FFT data that is pretty expensive to calculate, so that they are re-used for common sampling conversions.